refactor(web): fold detail routes into hooks, drop the -prefix hook naming hack
Two follow-ups to the earlier PR-review response:

1. journeys.$journeyId, lessons.$lessonId, and subjects.$subjectId still had
   the if(isPending)/if(error) early-return pattern inline in the component
   — only home.tsx had been fixed. Added useJourney/useLesson/useSubject
   hooks following the same discriminated-union shape as useHome, so all
   four detail routes are now thin switches with no business logic.

2. The dash-prefixed hook filenames (-useHome.ts) were a router-tooling
   escape hatch, not a real naming convention. Routes that need a hook now
   move from a flat <name>.tsx into <name>/index.tsx with the hook
   colocated as a plain useFile.ts alongside it (composition, not a name
   trick) — vite.config.ts's routeFileIgnorePattern excludes colocated
   hooks/tests from route generation instead. Routes with no hook
   (login/signup/onboarding/index) stay flat files.

Note: this changes each restructured route's canonical path to include a
trailing slash (e.g. /home -> /home/); the router still resolves the
no-slash form to the same route (verified via the existing test suite).
